requirements.txt
imutils~=0.5.4
opencv-python~=4.10.0.84
numpy~=1.26.4
pillow~=10.4.0
torch~=2.4.0
ultralytics~=8.2.82